Use string_view when splitting to reduce allocations.
This is quite an important change.  Thanks to std::string_view, we can
easily avoid allocating additional memory for the substrings when
splitting a tweet.  To account for this, the sentiment and cache map
keys are changed from std::string to unsigned long, the result of a
std::hash invocation.

Since a std::string and a std::string_view containing the same
characters and having the same length are guaranteed to produce the
same hash value (see
https://en.cppreference.com/w/cpp/string/basic_string/hash), there is
no danger of conflicts.  As long as the string being hashed contains
the same characters and has the same length as the one which was
originally hashed to build the maps, it will produce the same hash
value regardless of whether it’s a std::string or a std::string_view.

We could make this more "ergonomic" by overriding operator[], but
let’s not bother with it for now.  Explicitly using a std::hash object
also makes it more explicit that we are using a hashmap where keys
aren’t constrained by a specificy type.