Use the correct hashes in CachingClassifier.
For the result cache, we can map directly using strings as keys, since
we don’t need "compatibility" with string views.  Therefore, we just
use an (unordered_)map<string, SentimentResult> for the result cache,
and hash the tweet directly to provide potential cached results.  Of
course, if there is no cached result, we need to compute it on the
fly, as BasicClassifier always does.