Serialize total and handle empty samples case more gracefully.
We also serialize the total amount of latency measurements, including
unsampled ones.  For now, this is effectively equivalent to the amount
of tuples that were received by the sink(s).  We also handle the
degenerate case of an empty samples vector more gracefully: in this
unlikely scenario, we just set the mean every percentile to 0.0.  It’s
not informative at all, of course, but what can we expect with no
samples whatsoever?