Compute average latency correctly when printing out.
We must divide by the number of tuples actually sampled, not the total
number of tuples!  Of course, they are the same in case of a sampling
rate of 0 (sample every tuple), but this is not necessarily true, in
general.  This way, we can ensure that the values printed out to JSON
and stdout are identical (barring some very small rounding).