Use a dedicated Metric class to store metrics.
This is basically used as a wrapper for an "atomic" vector of samples
with a name.  The only "mutable" operation allowed on it is merging
samples from a std::vector, which are then promptly sorted.  An
invariant of this class is that individual samples are always sorted,
and this is ensured by the merge_samples() member function.  This
simplifies the interface to the rest of the program.