Use more consistent and intuitive type declarations
Sometimes, implicit conversion is actually desirable, especially for
built-in arithmetic types.  unsigned and long literals exist, but it’s
arguably more readable and intuitive to just name the type and assign
a plain integer literal to them.  "Uniform" initialization is not
really that uniform, it really only exists to avoid the most vexing parse.