Fix CircularFifoBuffer implementation
The problem was that the head index was updated right after inserting:
since the head index is also used by the get() member function, this
causes the function to not actually return the last inserted element,
rather, the one immediately after it (modulo the size of the buffer).

To fix this, we use an addtional "next_head" index which points to the
index that wil be used for the next insertion. "head" now points to
the last item that was actually inserted, if any.  As an invariant,
next_head is, before and after every "add" invocation, always equal
to (head + 1) % size, where size is the buffer size.