Emulate timers
In order to emulate timers in WindwFlow, we create an additional
source for each node in the graph that requires a timed event:  After
each period, the timer source sends a special "tick" tuple that
triggers the event:  this allows us to ensure that the timing of the
event is independent from the flow of "regular", non-tick tuples.

Of course, the disadvantage is that we need an additional thread for
each node with a timer, but since we only have three in this
application, the tradeoff seems to be worth it.