Add basic shared non-blocking queue implementation
If chaining is used, the queue might be pushed to and popped from the
same thread.  This means that if this thread is blocked on the
condition variable when popping, it will never be unblocked, since the
only thread that could unblock it is itself.

With this non-blocking version (which still uses locks, however), the
pop() member function returns an optional<T> instead of a T.  If the
queue is empty when popping, it returns a "null" value instead of
blocking: the caller can then decide what to do, including trying
again.  If the queue is not empty, pop a value and return it, as
usual.