Use NonBlockingQueue by default
The WindFlow runtime doesn’t seem to be able to wake up threads that
are blocked on a condition variable.  To work around this, we move the
wait "outside" the invocation to pop().

Note that the queue implementation itself is still not very efficient
because, while it does not block a thread on condition if it finds the
queue empty, it still uses a mutex, which must be acquired whenever we
attempt to pop the queue.  However, this shouldn’t be critical for our
purposes, and this simple change allows the WindFlow runtime to
correctly stop replicas when the computation is over.