Use an optional for parent timestamp
While it’s true that the value 0 for a timestamp is almost impossible,
it’s also still a valid numeric value that MIGHT be a valid
timestamp.  This would make the whole algorithm invalid.

A much better way is to use the std::optional class, provided by the
standard C++ library.  That way, we can better represent the case
where the variable contains an "invalid value".  With std::optinal,
the algorithm should be correct even in the unlikely case of a
timestamp equal to 0.