Do not mix ordering timestamps with observation timestamps.
Ordering timestamps are used to order tuples in case of a DEFAULT
execution mode.  However, the business logic relies on the timestamps
of the observations being non-decreasing, which are a distinct
concept.

In practice, we set the initial ordering timestamp with the current
observation timestmap, but then we update it with the current
timestamp value provided by the WindFlow RuntimeContext.  This allows
us to preserve the ordering while still using the same observation
timestamps.