Use secondary key for comparison in priority queues
Since it’s possible, albeit unlikely, to receive tuples with the same
ordering (WindFlow) timestamp, but with a different
observation (business logic) timestamp, we also sort according to the
observation timestamp as a secondary key.  This way, a tuples with the
same observation timestamps will still be processed in order,
according to the respective observation timestamp.