Add general implementation for TT
This implementation accepts a new command line argument,
--timernodes (-T), which determines whether to use additional WindFlow
nodes for timers, as in tt-timer-functors.cpp.  If we pass it a value
of 1 or the string "true", it uses additional timer nodes,
implementing timers with tick tuples.  Otherwise, it uses "internal"
threads like tt-timer-threads.cpp.  This allows us to have a single
source file, eliminating lots of duplication.