Set word field in tuple with a unique value per replica
Since some of the nodes with a timer have a KeyBy distribution, this
means that every tuple with the same key that the operator receives
will always be received by the same replica.  Unfortunately, this had
the unexpected side effect of effectively redirecting tick tuples to a
single replica, the one receiving the key matching the empty string.

This is because timer nodes did not explicitly initialize any field
except for is_tick_tuple, which is technically the only "useful" one.
However, this meant that the "word" field, which is an std::string,
was always initialized to the empty string, as per the default
constructor of std::string.

The solution is simply to assign a unique string per replica to the
word field, so that each tick tuple is always received by the intended
replica.  Notice that we don’t do this if the output type is a
RankingsTuple, since the node immediately after the Total Ranker (the
Sink) has no KeyBy distribution and hence we don’t need unique key
values.