Do not limit total parallelism degree
Using a sum of individual parallelism degrees that is higher than the
avaialble hardware concurrency can still be useful for debugging and
profiling purposes.  Hence, we remove these checks.  Whoever runs the
application is required to ensure that the total parallelism degree
does not exceed the available number of hardware threads when doing a
production or performance testing run.