Do not process last tuples if there is no data in local storage
Due to the keyby distribution, some replicas may never run their
operator() at all, which means that the respective localstorage object
won’t contain a data object.  This means that we must check if it
actually exists in storage before processing leftover tuples.  If the
data was never allocated, we do nothing.